moves types to separate files, fixes type imports
