+++ src/store/action.ts
@@ -1,6 +1,6 @@
 import {createAction} from '@reduxjs/toolkit';
 import {AuthorizationStatusModel, SixCitiesModel} from '../const/const.ts';
-import {FullOfferModel, OfferModel,} from '../types/types.ts';
+import {FullOfferModel, OfferModel,} from '../types/offer-model.ts';
 import {ReviewModel} from '../types/review-model.ts';
 
 export const selectCity = createAction<SixCitiesModel>('main/selectCity');
