+++ src/store/api-actions.ts
@@ -1,11 +1,11 @@
 import {AxiosInstance} from 'axios';
 import {createAsyncThunk} from '@reduxjs/toolkit';
 import {AppDispatch, State} from './index.ts';
-import {FullOfferModel, OfferModel,} from '../types/types.ts';
-import {APIRoute, AuthorizationStatus} from '../const/const.ts';
-import {UserData,} from '../types/user-data.ts';
+import {FullOfferModel, OfferModel} from '../types/offer-model.ts';
+import {AuthorizationStatus, Endpoint} from '../const/const.ts';
+import {FullUserModel,} from '../types/user-model.ts';
 import {dropToken, saveToken} from '../services/token.ts';
-import {AuthData} from '../types/auth-data.ts';
+import {AuthModel} from '../types/auth-model.ts';
 import {
   loadFavorites,
   loadNearby,
@@ -30,7 +30,7 @@ export const fetchOffersAction = createAsyncThunk<void, undefined, {
   'data/fetchOffers',
   async (_arg, {dispatch, extra: api}) => {
     dispatch(setOffersDataLoadingStatus(true));
-    const {data} = await api.get<OfferModel[]>(APIRoute.Offers);
+    const {data} = await api.get<OfferModel[]>(Endpoint.Offers);
     dispatch(setOffersDataLoadingStatus(false));
     dispatch(loadOffers(data));
   },
@@ -44,7 +44,7 @@ export const fetchFavoritesAction = createAsyncThunk<void, undefined, {
   'data/fetchFavorites',
   async (_arg, {dispatch, extra: api}) => {
     dispatch(setFavoritesDataLoadingStatus(true));
-    const {data} = await api.get<OfferModel[]>(APIRoute.Favorite);
+    const {data} = await api.get<OfferModel[]>(Endpoint.Favorite);
     dispatch(loadFavorites(data));
     dispatch(setFavoritesDataLoadingStatus(false));
   },
@@ -57,7 +57,7 @@ export const fetchOfferByIdAction = createAsyncThunk<void, Record<'id', string>,
 }>(
   'data/fetchOfferById',
   async ({id}, {dispatch, extra: api}) => {
-    const {data} = await api.get<FullOfferModel>(`${APIRoute.Offers}/${id}`);
+    const {data} = await api.get<FullOfferModel>(`${Endpoint.Offers}/${id}`);
     dispatch(loadOfferById(data));
   },
 );
@@ -69,7 +69,7 @@ export const fetchNearbyAction = createAsyncThunk<void, Record<'id', string>, {
 }>(
   'data/fetchNearbyAction',
   async ({id}, {dispatch, extra: api}) => {
-    const {data} = await api.get<OfferModel[]>(`${APIRoute.Offers}/${id}/nearby`);
+    const {data} = await api.get<OfferModel[]>(`${Endpoint.Offers}/${id}/nearby`);
     dispatch(loadNearby(data));
   },
 );
@@ -81,7 +81,7 @@ export const fetchReviewsAction = createAsyncThunk<void, Record<'id', string>, {
 }>(
   'data/loadReviews',
   async ({id}, {dispatch, extra: api}) => {
-    const {data} = await api.get<ReviewModel[]>(`${APIRoute.Reviews}/${id}`);
+    const {data} = await api.get<ReviewModel[]>(`${Endpoint.Reviews}/${id}`);
     dispatch(loadReviews(data));
   },
 );
@@ -94,7 +94,7 @@ export const sendReviewAction = createAsyncThunk<void, NewReviewModel, {
   'data/sendReview',
   async ({id, comment, rating}, {dispatch, extra: api}) => {
     try {
-      await api.post<ReviewModel>(`${APIRoute.Reviews}/${id}`, {comment, rating});
+      await api.post<ReviewModel>(`${Endpoint.Reviews}/${id}`, {comment, rating});
       dispatch(fetchReviewsAction({id}));
       dispatch(setCommentSendStatus(true));
     } catch {
@@ -111,7 +111,7 @@ export const checkAuthAction = createAsyncThunk<void, undefined, {
   'user/checkAuth',
   async (_arg, {dispatch, extra: api}) => {
     try {
-      const {data: {email}} = await api.get<UserData>(APIRoute.Login);
+      const {data: {email}} = await api.get<FullUserModel>(Endpoint.Login);
       dispatch(requireAuthorization(AuthorizationStatus.Auth));
       dispatch(saveUserEmail(email));
     } catch {
@@ -120,14 +120,14 @@ export const checkAuthAction = createAsyncThunk<void, undefined, {
   },
 );
 
-export const loginAction = createAsyncThunk<void, AuthData, {
+export const loginAction = createAsyncThunk<void, AuthModel, {
   dispatch: AppDispatch;
   state: State;
   extra: AxiosInstance;
 }>(
   'user/login',
   async ({email, password}, {dispatch, extra: api}) => {
-    const {data: {token}} = await api.post<UserData>(APIRoute.Login, {email, password});
+    const {data: {token}} = await api.post<FullUserModel>(Endpoint.Login, {email, password});
     saveToken(token);
     dispatch(requireAuthorization(AuthorizationStatus.Auth));
     dispatch(saveUserEmail(email));
@@ -141,7 +141,7 @@ export const logoutAction = createAsyncThunk<void, undefined, {
 }>(
   'user/logout',
   async (_arg, {dispatch, extra: api}) => {
-    await api.delete(APIRoute.Logout);
+    await api.delete(Endpoint.Logout);
     dropToken();
     dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
   },
