+++ src/store/reducer.ts
@@ -13,7 +13,7 @@ import {
   setOffersDataLoadingStatus
 } from './action.ts';
 import {AuthorizationStatus, AuthorizationStatusModel, SixCitiesModel} from '../const/const.ts';
-import {FullOfferModel, OfferModel} from '../types/types.ts';
+import {FullOfferModel, OfferModel} from '../types/offer-model.ts';
 import {ReviewModel} from '../types/review-model.ts';
 
 export type InitialStateModel = {
