+++ src/types/city-model.ts
@@ -1,0 +1,7 @@
+import {SixCitiesModel} from '../const/const.ts';
+import {LocationModel} from './location-model.ts';
+
+export type CityModel = {
+  name: SixCitiesModel;
+  location: LocationModel;
+}
