+++ src/types/location-model.ts
@@ -1,0 +1,5 @@
+export type LocationModel = {
+  latitude: number;
+  longitude: number;
+  zoom: number;
+}
