+++ src/types/place-card-model.ts
