+++ src/types/point-model.ts
@@ -1,4 +1,4 @@
-import {LocationModel} from './types.ts';
+import {LocationModel} from './location-model.ts';
 
 export type PointModel = {
   id: string;
