--- src/types/user-data.ts
@@ -1,7 +1,0 @@
-export type UserData = {
-  name: string;
-  avatarUrl: string;
-  isPro: false;
-  email: string;
-  token: string;
-};
