+++ src/types/user-model.ts
@@ -1,0 +1,10 @@
+export type UserModel = {
+  name: string;
+  avatarUrl: string;
+  isPro: boolean;
+}
+
+export type FullUserModel = UserModel & {
+  email: string;
+  token: string;
+};
